// src/types/Builder.ts
export type FieldValue = number | string | Record<string, any> | undefined | null | boolean;

export interface IRecord {
    [$key: string]: FieldValue;
}
// ...
export interface IBuilderConfig extends IBuilderBaseConfig {
    // Идентификатор класса
    ck_id?: string;
    // Ротельно класса
    ck_parent: string;
    ck_master?: string;
    ck_query?: string;
    ck_page_object?: string;
    cv_displayed?: string;
    cv_name?: string;
    ck_object?: string;
    cn_order?: number;
    // Определяет класс в качестве мастера
    cl_is_master?: boolean;
    contentwidth?: string;
    // Вложенные элементы для построения контекстного меню
    contextmenus?: IBuilderConfig[];
    // Служебный параметр для иерархии
    bottombtn?: IBuilderConfig[];
    // Служебный параметр для иерархии
    childs?: IBuilderConfig[];
    // Служебный параметр для иерархии
    childwindow?: IBuilderConfig[];
    // Служебный параметр для иерархии
    columns?: IBuilderConfig[];
    // Служебный параметр для иерархии
    editors?: IBuilderConfig[] | 'false';
    // Служебный параметр для иерархии
    filters?: IBuilderConfig[];
    // Служебный параметр для иерархии
    topbtn?: IBuilderConfig[];
    // Отображение панели с детализацией - Служебная информация
    detail?: IBuilderConfig[];
    // Служебный параметр для передачи статических значений
    records?: Record<string, FieldValue>[];
    // Служебный параметр не править
    type: string;
    // Признак Только чтение
    readonly?: boolean;
    // Interanal
    confirmquestionposition?: 'right' | 'top';
    // Internal
    iconsize?: 'xs';
    // Internal values for window store or etc.
    values?: IRecord;
    // Internal Возможность выделения нескольких значений в GRID и TREEGRID SINGLE - только 1 значение SIMPLE - позволяет выбирать значения одно-за-другим. Каждое нажатие добавляет/удаляет значение. MULTI - позволяет комплексно выбирать значения, с учетом ctrl и shift
    selmode?: 'MULTI';
}